[rename] InteriorBillDto 패키지 변경
