+++ backend/src/main/java/com/autoever/idle/domain/interiorColor/InteriorColorRepositoryImpl.java
@@ -1,5 +1,6 @@
 package com.autoever.idle.domain.interiorColor;
 
+import com.autoever.idle.domain.interiorColor.dto.InteriorBillDto;
 import com.autoever.idle.domain.interiorColor.dto.InteriorColorDto;
 import org.springframework.jdbc.core.BeanPropertyRowMapper;
 import org.springframework.jdbc.core.JdbcTemplate;
